refactor: Organize repository files into a dedicated directory
