+++ src/category/category.modules.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { CategoryService } from "./category.service";
 import { CategoryController } from "./category.controller";
-import { CategoryRepository } from "./category.repository";
+import { CategoryRepository } from "./repository/category.repository";
 import { PrismaService } from "src/database/prisma.service";
 
 @Module({
