+++ src/item/repository/item.repository.ts
@@ -1,0 +1,2 @@
+import { Injectable } from "@nestjs/common";
+import { ItemService } from "./item.service";
